feat: add answers to prepolation
