+++ apps/ehr/src/helpers/fhir.ts
@@ -1,7 +1,6 @@
 import { BatchInputRequest } from '@oystehr/sdk';
 import { Operation } from 'fast-json-patch';
-import { FhirResource, Patient } from 'fhir/r4b';
-import { PATIENT_INDIVIDUAL_PRONOUNS_URL } from 'utils';
+import { FhirResource } from 'fhir/r4b';
 
 export interface GetPatchBinaryInput {
   resourceId: string;
@@ -21,12 +20,3 @@ export function getPatchBinary<F extends FhirResource>(input: GetPatchBinaryInpu
     },
   };
 }
-
-// helper function to get pronouns display value from extension
-export const getPronounsFromExtension = (patient: Patient): string => {
-  const pronounsExtension = patient.extension?.find(
-    (ext: { url: string }) => ext.url === PATIENT_INDIVIDUAL_PRONOUNS_URL
-  );
-  if (!pronounsExtension?.valueCodeableConcept?.coding?.[0]) return '';
-  return pronounsExtension.valueCodeableConcept.coding[0].display || '';
-};
